move JobCollection, JobScheduler, JobSchedulerFactory and  relevant tests into their own files
