+++ tron/api/controller.py
@@ -1,9 +1,6 @@
 """
 Web Controllers for the API.
 """
-from __future__ import absolute_import
-from __future__ import unicode_literals
-
 import logging
 
 from tron import yaml
